[ResponseOps] Prepare the connector `create` HTTP API for versioning (#194879)
Towards https://github.com/elastic/response-ops-team/issues/125

## Summary

Preparing the `POST ${BASE_ACTION_API_PATH}/connector/{id?}` HTTP API
for versioning

(cherry picked from commit 5adb0ea5d2c526aa4382677a508e1be0058bf890)
