+++ x-pack/plugins/actions/common/routes/connector/apis/create/schemas/latest.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export { transformGetConnectorResponse } from './v1';
+export * from './v1';
