+++ x-pack/plugins/actions/server/actions_client/actions_client.ts
@@ -14,9 +14,7 @@ import { compact, uniq } from 'lodash';
 import {
   IScopedClusterClient,
   SavedObjectsClientContract,
-  SavedObjectAttributes,
   KibanaRequest,
-  SavedObjectsUtils,
   Logger,
 } from '@kbn/core/server';
 import { AuditLogger } from '@kbn/security-plugin/server';
@@ -29,6 +27,7 @@ import { get } from '../application/connector/methods/get';
 import { getAll, getAllSystemConnectors } from '../application/connector/methods/get_all';
 import { update } from '../application/connector/methods/update';
 import { listTypes } from '../application/connector/methods/list_types';
+import { create } from '../application/connector/methods/create';
 import { execute } from '../application/connector/methods/execute';
 import {
   GetGlobalExecutionKPIParams,
@@ -36,15 +35,7 @@ import {
   IExecutionLogResult,
 } from '../../common';
 import { ActionTypeRegistry } from '../action_type_registry';
-import {
-  validateConfig,
-  validateSecrets,
-  ActionExecutorContract,
-  validateConnector,
-  ActionExecutionSource,
-  parseDate,
-  tryCatch,
-} from '../lib';
+import { ActionExecutorContract, ActionExecutionSource, parseDate } from '../lib';
 import {
   ActionResult,
   RawAction,
@@ -94,19 +85,10 @@ import {
 import { connectorFromSavedObject, isConnectorDeprecated } from '../application/connector/lib';
 import { ListTypesParams } from '../application/connector/methods/list_types/types';
 import { ConnectorUpdateParams } from '../application/connector/methods/update/types';
-import { ConnectorUpdate } from '../application/connector/methods/update/types/types';
+import { ConnectorCreateParams } from '../application/connector/methods/create/types';
 import { isPreconfigured } from '../lib/is_preconfigured';
 import { isSystemAction } from '../lib/is_system_action';
 import { ConnectorExecuteParams } from '../application/connector/methods/execute/types';
-
-interface Action extends ConnectorUpdate {
-  actionTypeId: string;
-}
-
-export interface CreateOptions {
-  action: Action;
-  options?: { id?: string };
-}
 
 export interface ConstructorOptions {
   logger: Logger;
@@ -187,156 +169,10 @@ export class ActionsClient {
    * Create an action
    */
   public async create({
-    action: { actionTypeId, name, config, secrets },
+    action,
     options,
-  }: CreateOptions): Promise<ActionResult> {
-    const id = options?.id || SavedObjectsUtils.generateId();
-
-    try {
-      await this.context.authorization.ensureAuthorized({
-        operation: 'create',
-        actionTypeId,
-      });
-    } catch (error) {
-      this.context.auditLogger?.log(
-        connectorAuditEvent({
-          action: ConnectorAuditAction.CREATE,
-          savedObject: { type: 'action', id },
-          error,
-        })
-      );
-      throw error;
-    }
-
-    const foundInMemoryConnector = this.context.inMemoryConnectors.find(
-      (connector) => connector.id === id
-    );
-
-    if (
-      this.context.actionTypeRegistry.isSystemActionType(actionTypeId) ||
-      foundInMemoryConnector?.isSystemAction
-    ) {
-      throw Boom.badRequest(
-        i18n.translate('xpack.actions.serverSideErrors.systemActionCreationForbidden', {
-          defaultMessage: 'System action creation is forbidden. Action type: {actionTypeId}.',
-          values: {
-            actionTypeId,
-          },
-        })
-      );
-    }
-
-    if (foundInMemoryConnector?.isPreconfigured) {
-      throw Boom.badRequest(
-        i18n.translate('xpack.actions.serverSideErrors.predefinedIdConnectorAlreadyExists', {
-          defaultMessage: 'This {id} already exists in a preconfigured action.',
-          values: {
-            id,
-          },
-        })
-      );
-    }
-
-    const actionType = this.context.actionTypeRegistry.get(actionTypeId);
-    const configurationUtilities = this.context.actionTypeRegistry.getUtils();
-    const validatedActionTypeConfig = validateConfig(actionType, config, {
-      configurationUtilities,
-    });
-    const validatedActionTypeSecrets = validateSecrets(actionType, secrets, {
-      configurationUtilities,
-    });
-    if (actionType.validate?.connector) {
-      validateConnector(actionType, { config, secrets });
-    }
-    this.context.actionTypeRegistry.ensureActionTypeEnabled(actionTypeId);
-
-    const hookServices: HookServices = {
-      scopedClusterClient: this.context.scopedClusterClient,
-    };
-
-    if (actionType.preSaveHook) {
-      try {
-        await actionType.preSaveHook({
-          connectorId: id,
-          config,
-          secrets,
-          logger: this.context.logger,
-          request: this.context.request,
-          services: hookServices,
-          isUpdate: false,
-        });
-      } catch (error) {
-        this.context.auditLogger?.log(
-          connectorAuditEvent({
-            action: ConnectorAuditAction.CREATE,
-            savedObject: { type: 'action', id },
-            error,
-          })
-        );
-        throw error;
-      }
-    }
-
-    this.context.auditLogger?.log(
-      connectorAuditEvent({
-        action: ConnectorAuditAction.CREATE,
-        savedObject: { type: 'action', id },
-        outcome: 'unknown',
-      })
-    );
-
-    const result = await tryCatch(
-      async () =>
-        await this.context.unsecuredSavedObjectsClient.create(
-          'action',
-          {
-            actionTypeId,
-            name,
-            isMissingSecrets: false,
-            config: validatedActionTypeConfig as SavedObjectAttributes,
-            secrets: validatedActionTypeSecrets as SavedObjectAttributes,
-          },
-          { id }
-        )
-    );
-
-    const wasSuccessful = !(result instanceof Error);
-    const label = `connectorId: "${id}"; type: ${actionTypeId}`;
-    const tags = ['post-save-hook', id];
-
-    if (actionType.postSaveHook) {
-      try {
-        await actionType.postSaveHook({
-          connectorId: id,
-          config,
-          secrets,
-          logger: this.context.logger,
-          request: this.context.request,
-          services: hookServices,
-          isUpdate: false,
-          wasSuccessful,
-        });
-      } catch (err) {
-        this.context.logger.error(`postSaveHook create error for ${label}: ${err.message}`, {
-          tags,
-        });
-      }
-    }
-
-    if (!wasSuccessful) {
-      throw result;
-    }
-
-    return {
-      id: result.id,
-      actionTypeId: result.attributes.actionTypeId,
-      isMissingSecrets: result.attributes.isMissingSecrets,
-      name: result.attributes.name,
-      config: result.attributes.config,
-      isPreconfigured: false,
-      isSystemAction: false,
-      isDeprecated: isConnectorDeprecated(result.attributes),
-    };
+  }: Omit<ConnectorCreateParams, 'context'>): Promise<ActionResult> {
+    return create({ context: this.context, action, options });
   }
 
   /**
