+++ x-pack/plugins/actions/server/routes/connector/create/transforms/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { transformCreateConnectorBody } from './transform_connector_body/latest';
+
+export { transformCreateConnectorBody as transformCreateConnectorBodyV1 } from './transform_connector_body/v1';
