+++ x-pack/plugins/actions/server/routes/index.ts
@@ -13,7 +13,7 @@ import { listTypesRoute } from './connector/list_types';
 import { listTypesWithSystemRoute } from './connector/list_types_system';
 import { ILicenseState } from '../lib';
 import { ActionsRequestHandlerContext } from '../types';
-import { createActionRoute } from './create';
+import { createConnectorRoute } from './connector/create';
 import { deleteConnectorRoute } from './connector/delete';
 import { executeConnectorRoute } from './connector/execute';
 import { getConnectorRoute } from './connector/get';
@@ -36,7 +36,7 @@ export function defineRoutes(opts: RouteOptions) {
 
   defineLegacyRoutes(router, licenseState, usageCounter);
 
-  createActionRoute(router, licenseState);
+  createConnectorRoute(router, licenseState);
   deleteConnectorRoute(router, licenseState);
   getConnectorRoute(router, licenseState);
   getAllConnectorsRoute(router, licenseState);
