+++ x-pack/plugins/security_solution/scripts/endpoint/common/connectors_services.ts
@@ -7,8 +7,7 @@
 
 import type { KbnClient } from '@kbn/test';
 import type { AllConnectorsResponseV1 } from '@kbn/actions-plugin/common/routes/connector/response';
-import type { bodySchema } from '@kbn/actions-plugin/server/routes/create';
-import type { TypeOf } from '@kbn/config-schema';
+import type { CreateConnectorRequestBodyV1 } from '@kbn/actions-plugin/common/routes/connector/apis/create';
 import type { Connector } from '@kbn/actions-plugin/server/application/connector/types';
 import { catchAxiosErrorFormatAndThrow } from '../../../common/endpoint/format_axios_error';
 
@@ -46,8 +45,6 @@ export const fetchConnectorByType = async (
   }
 };
 
-type CreateConnectorBody = TypeOf<typeof bodySchema>;
-
 /**
  * Creates a connector in the stack
  * @param kbnClient
@@ -55,7 +52,7 @@ type CreateConnectorBody = TypeOf<typeof bodySchema>;
  */
 export const createConnector = async (
   kbnClient: KbnClient,
-  createPayload: CreateConnectorBody
+  createPayload: CreateConnectorRequestBodyV1
 ): Promise<Connector> => {
   return kbnClient
     .request<Connector>({
